requirements.txt
# Google Sheets API
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.128.0

# OpenAI
openai>=1.12.0

# Gmail API
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# Queue and workers
redis>=5.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Email handling
email-validator>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Async
asyncio
aiohttp>=3.9.0
aiofiles>=23.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
dateutil>=2.8.0
pytz>=2024.1

# Logging
structlog>=24.1.0

# For testing
pytest>=7.4.0
pytest-asyncio>=0.23.0